requirements.txt
pyautogen>=0.2.32
openai>=1.30.0
python-dotenv>=1.0.1
yfinance>=0.2.40
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0